Back-end data collection and "Other" text-box
Selection of the type of payment and amount to be payed will store the respective values in a hidden textbox which will send the data to backend when clicked Donate Now. But no path or method have been added for backend.
The "Other" box is now a text box.
Also, some slight modifications for screen fitting.